apps/suggestions/requirements/requirements.txt
Flask==2.0.3
black==22.6.0
dependency-injector==4.40.0
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
protobuf~=3.20.0
